requirements.txt
openai
mdpdf
streamlit
watchdog
google-auth-oauthlib
google-auth-httplib2
google-api-python-client